refactor parse submodules into individual parser files
