+++ src/parse/meta/not_parser.rs
@@ -1,0 +1,15 @@
+use super::super::Token;
+use super::and_parser::parse_and;
+use crate::{Error, Pattern, Result};
+
+pub(crate) fn parse_not(lexer: &mut logos::Lexer<Token>) -> Result<Pattern> {
+    let mut lookahead = lexer.clone();
+    match lookahead.next() {
+        Some(Ok(Token::Not)) => {
+            lexer.next();
+            let pat = parse_not(lexer)?;
+            Ok(Pattern::not_matching(pat))
+        }
+        _ => parse_and(lexer),
+    }
+}
