+++ src/parse/structure/assertion_parser.rs
@@ -1,0 +1,8 @@
+use super::super::Token;
+use crate::{Pattern, Result};
+
+pub(crate) fn parse_assertion(
+    _lexer: &mut logos::Lexer<Token>,
+) -> Result<Pattern> {
+    Ok(Pattern::any_assertion())
+}
