+++ src/parse/structure/predicate_parser.rs
@@ -1,0 +1,23 @@
+use super::super::{Token, meta};
+use crate::{Error, Pattern, Result};
+
+pub(crate) fn parse_predicate(
+    lexer: &mut logos::Lexer<Token>,
+) -> Result<Pattern> {
+    let mut lookahead = lexer.clone();
+    match lookahead.next() {
+        Some(Ok(Token::ParenOpen)) => {
+            lexer.next();
+            let pat = meta::parse_or(lexer)?;
+            match lexer.next() {
+                Some(Ok(Token::ParenClose)) => Ok(Pattern::predicate(pat)),
+                Some(Ok(t)) => {
+                    Err(Error::UnexpectedToken(Box::new(t), lexer.span()))
+                }
+                Some(Err(e)) => Err(e),
+                None => Err(Error::ExpectedCloseParen(lexer.span())),
+            }
+        }
+        _ => Ok(Pattern::any_predicate()),
+    }
+}
